+++ ChangeLog.md
@@ -1,4 +1,6 @@
 # HorizonsBot Change Log
+#### HorizonsBot Version 2.0.2:
+- Fixed club events being scheduled for the current meeting instead of the next meeting
 #### HorizonsBot Version 2.0.1:
 - Clubs that aren't recruiting are filtered out of the club list select
 - Fixed selects (drop downs) crashing the bot